backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.1
psycopg2-binary==2.9.10
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.20
aiosmtplib==3.0.2
email-validator==2.2.0
geoalchemy2==0.15.2
httpx==0.28.1
